refactor: change guards name and delete unused imports & logs
