+++ src/modules/users/users.schema.ts
@@ -1,4 +1,4 @@
-import { relations, sql } from 'drizzle-orm'
+import { relations } from 'drizzle-orm'
 import {
   pgEnum,
   pgTable,
